CSV file for resolving overlapping disease categories in mapped CALIBER codes

Returns the file path to csv included with this package for resolving overlapping disease categories in mapped CALIBER codes. To be used with reformat_caliber_for_ukb().

Usage

default_overlapping_disease_categories_csv()

Value

A file path.

See also

Examples

# return file path
default_overlapping_disease_categories_csv()
#> [1] "/home/runner/work/_temp/Library/codeminer/extdata/caliber_mapped_overlapping_disease_categories.csv"

# read file into R
read.csv(default_overlapping_disease_categories_csv())
#>                                disease
#> 1                             Diabetes
#> 2                             Diabetes
#> 3  Diabetic neurological complications
#> 4  Diabetic neurological complications
#> 5  Diabetic neurological complications
#> 6  Diabetic neurological complications
#> 7              End stage renal disease
#> 8              End stage renal disease
#> 9                 Erectile dysfunction
#> 10                Erectile dysfunction
#> 11     Primary Malignancy_Other Organs
#> 12     Primary Malignancy_Other Organs
#> 13     Primary Malignancy_Other Organs
#> 14     Primary Malignancy_Other Organs
#> 15     Primary Malignancy_Other Organs
#> 16     Primary Malignancy_Other Organs
#> 17     Primary Malignancy_Other Organs
#> 18     Primary Malignancy_Other Organs
#> 19     Primary Malignancy_Other Organs
#> 20     Primary Malignancy_Other Organs
#> 21     Primary Malignancy_Other Organs
#> 22     Primary Malignancy_Other Organs
#> 23     Primary Malignancy_Other Organs
#> 24     Primary Malignancy_Other Organs
#> 25     Primary Malignancy_Other Organs
#> 26     Primary Malignancy_Other Organs
#> 27     Primary Malignancy_Other Organs
#> 28     Primary Malignancy_Other Organs
#> 29                        Tuberculosis
#> 30                        Tuberculosis
#> 31                     Unstable Angina
#> 32                     Unstable Angina
#>                                                  description
#> 1                                          DIABETES MELLITUS
#> 2                                          DIABETES MELLITUS
#> 3  Diabetic (femoral mononeuropathy) & (Diabetic amyotrophy)
#> 4  Diabetic (femoral mononeuropathy) & (Diabetic amyotrophy)
#> 5                                        Diabetic amyotrophy
#> 6                                        Diabetic amyotrophy
#> 7                                    End stage renal failure
#> 8                                    End stage renal failure
#> 9                                       Erectile dysfunction
#> 10                                      Erectile dysfunction
#> 11                                       HEAD, FACE AND NECK
#> 12                                       HEAD, FACE AND NECK
#> 13                            UPPER LIMB, INCLUDING SHOULDER
#> 14                            UPPER LIMB, INCLUDING SHOULDER
#> 15                                 LOWER LIMB, INCLUDING HIP
#> 16                                 LOWER LIMB, INCLUDING HIP
#> 17                                                    THORAX
#> 18                                                    THORAX
#> 19                                                   ABDOMEN
#> 20                                                   ABDOMEN
#> 21                                                    PELVIS
#> 22                                                    PELVIS
#> 23                                        TRUNK, UNSPECIFIED
#> 24                                        TRUNK, UNSPECIFIED
#> 25                                                     OTHER
#> 26                                                     OTHER
#> 27                                                     OTHER
#> 28                                                     OTHER
#> 29          Late effects of tuberculosis of bones and joints
#> 30          Late effects of tuberculosis of bones and joints
#> 31                                          Worsening angina
#> 32                                          Worsening angina
#>                                                 category code_type  code
#> 1                   Diabetes not otherwise specified (6)      icd9  6480
#> 2                                 Secondary diabetes (5)      icd9  6480
#> 3       Diagnosis of diabetic neurological complications     read3 Xa0lK
#> 4         History of diabetic neurological complications     read3 Xa0lK
#> 5       Diagnosis of diabetic neurological complications     read3 XaPmX
#> 6         History of diabetic neurological complications     read3 XaPmX
#> 7                   Diagnosis of End stage renal disease     read3 X30J0
#> 8                  Procedure for End stage renal disease     read3 X30J0
#> 9                      Diagnosis of erectile dysfunction     read3 E2273
#> 10            Possible diagnosis of erectile dysfunction     read3 E2273
#> 11          Diagnosis of Primary Malignancy_Other Organs      icd9  1710
#> 12 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1710
#> 13          Diagnosis of Primary Malignancy_Other Organs      icd9  1712
#> 14 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1712
#> 15          Diagnosis of Primary Malignancy_Other Organs      icd9  1713
#> 16 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1713
#> 17          Diagnosis of Primary Malignancy_Other Organs      icd9  1714
#> 18 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1714
#> 19          Diagnosis of Primary Malignancy_Other Organs      icd9  1715
#> 20 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1715
#> 21          Diagnosis of Primary Malignancy_Other Organs      icd9  1716
#> 22 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1716
#> 23          Diagnosis of Primary Malignancy_Other Organs      icd9  1717
#> 24 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1717
#> 25          Diagnosis of Primary Malignancy_Other Organs      icd9  1718
#> 26 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1718
#> 27          Diagnosis of Primary Malignancy_Other Organs      icd9  1878
#> 28 Possible Diagnosis of Primary Malignancy_Other Organs      icd9  1878
#> 29                             Diagnosis of tuberculosis     read3 AE03.
#> 30                               History of tuberculosis     read3 AE03.
#> 31                                   Unstable angina (3)     read3 XE0Ui
#> 32                                  Worsening angina (2)     read3 XE0Ui
#>     author keep
#> 1  caliber    Y
#> 2  caliber     
#> 3  caliber    Y
#> 4  caliber     
#> 5  caliber    Y
#> 6  caliber     
#> 7  caliber    Y
#> 8  caliber     
#> 9  caliber    Y
#> 10 caliber     
#> 11 caliber    Y
#> 12 caliber     
#> 13 caliber    Y
#> 14 caliber     
#> 15 caliber    Y
#> 16 caliber     
#> 17 caliber    Y
#> 18 caliber     
#> 19 caliber    Y
#> 20 caliber     
#> 21 caliber    Y
#> 22 caliber     
#> 23 caliber    Y
#> 24 caliber     
#> 25 caliber    Y
#> 26 caliber     
#> 27 caliber    Y
#> 28 caliber     
#> 29 caliber    Y
#> 30 caliber     
#> 31 caliber     
#> 32 caliber    Y