Get parents for SNOMED codes

Get parents for SNOMED codes

Usage

get_parents_sct(
  codes,
  standardise_output = TRUE,
  include_self = TRUE,
  include_ancestors = TRUE,
  all_lkps_maps = NULL,
  preferred_description_only = TRUE,
  col_filters = getOption("codeminer.col_filters")
)

Arguments

codes

Character vector of SNOMED codes.

standardise_output

If TRUE (default) return a data frame with columns 'code', 'description' and 'code_type'.

include_self

If TRUE (default) include input codes in the result.

include_ancestors

If TRUE (default) return all ancestor codes, as well as immediate parents.

all_lkps_maps

Either a named list of lookup and mapping tables (either data frames or tbl_dbi objects), or the path to a Duckdb database containing these tables. If NULL, will attempt to connect to a Duckdb database named 'all_lkps_maps.db' in the current working directory, or to a a Duckdb database specified by an environmental variable named 'ALL_LKPS_MAPS_DB' (see here for how to set environment variables using a .Renviron file). The latter method will be used in preference.

preferred_description_only

If TRUE (default), return only preferred descriptions for clinical codes with synonyms. Will only apply if standardise_output is also TRUE.

col_filters

A named list where each name in the list refers to the name of a lookup or mapping table. Each item is also a named list, where the names refer to column names in the corresponding table, and the items are vectors of values to filter for. For example, list(my_lookup_table = list(colA = c("A", "B")) will result in my_lookup_table being filtered for rows where colA is either 'A' or 'B'. Uses default_col_filters() by default. Set to NULL to remove all filters.

Value

A dataframe

See also