Map clinical codes from one coding system to another

Map clinical codes from one coding system to another

Usage

MAP(
  codes,
  from = getOption("codeminer.map_from"),
  to = getOption("codeminer.map_to"),
  version = getOption("codeminer.map_version", default = "latest")
)

Arguments

codes

A character vector of codes to be mapped. If passing "all", all mapped codes will be returned.

from

Coding system that codes belong to.

to

Coding system to map codes to.

version

Version of the mapping table to use.

Value

A data.frame of the mapped codes with their descriptions, as returned by CODES().

If using codes = "all", returns the mapping table as a data.frame with columns:

from: the codes from the source code system

to: the mapped codes from the destination system

Details

If no mapping table matching the from -> to direction is found, but there is a table for to -> from, MAP() will return the reverse mapping with a warning. Note that this is not guaranteed to be correct, as most mapping tables only work one way.

See also

Examples

# Set up a temporary dummy database
temp_db <- tempfile(fileext = ".duckdb")
create_dummy_database(temp_db)
#> Creating new database at /tmp/RtmpIG4lZV/file19106282278b.duckdb
#> ✔ Lookup table icd10_v0 added successfully.
#> ✔ Lookup table read3_v0 added successfully.
#> ✔ Mapping table read3_icd10_v0 added successfully.
#> ✔ Dummy database ready to use!

MAP("X40J4", from = "read3", to = "icd10")
#> Warning: cannot open file '/home/runner/.local/share/codeminer/ontology.duckdb': No such file or directory
#> Error in file(con, "w"): cannot open the connection

# Return the mapping table itself
MAP("all", from = "read3", to = "icd10")
#> Warning: cannot open file '/home/runner/.local/share/codeminer/ontology.duckdb': No such file or directory
#> Error in file(con, "w"): cannot open the connection